message

+++ pages/4_Project-overview.py
@@ -465,7 +465,7 @@ def main():
 
         df5 = df5.dropna()
 
-        df5['Total_cost'] = df5['duration'] * df5['value']
+        df5['Total_cost'] = (df5['duration'].astype(int)) * (df5['value'].astype(int))
 
         st.write("After Preprocessing Data from Query",df5)
         # df5_filtered = df5[df5['year']==selected_option]
